requirements.txt
gradio
ultralytics
psycopg2-binary
google-genai
sentence-transformers
python-dotenv
markdown